split identifiers into separate files
